migrate char database to EF Core * cleanup * removed autocomplete as it might case problems when something goes wrong (UoW will be disposed and changes will still be safed)
