+++ src/Rasa.DBL/Repositories/IUnitOfWork.cs
@@ -4,7 +4,6 @@ namespace Rasa.Repositories
 {
     public interface IUnitOfWork : IDisposable
     {
-        bool AutoComplete { get; set; }
         void Complete();
         void Reject();
     }
