+++ src/Rasa.DBL/Repositories/IUnitOfWorkFactory.cs
@@ -1,0 +1,9 @@
+﻿namespace Rasa.Repositories
+{
+    using Char;
+
+    public interface IUnitOfWorkFactory
+    {
+        ICharUnitOfWork CreateCharUnitOfWork();
+    }
+}
