+++ src/Rasa.DBL/Repositories/UnitOfWork.cs
@@ -11,8 +11,6 @@ namespace Rasa.Repositories
             _dbContext = dbContext;
         }
 
-        public bool AutoComplete { get; set; }
-
         public void Complete()
         {
             if (_dbContext.ChangeTracker.HasChanges())
@@ -31,10 +29,6 @@ namespace Rasa.Repositories
 
         public void Dispose()
         {
-            if (AutoComplete)
-            {
-                Complete();
-            }
             Reject();
             _dbContext.Dispose();
         }
